Moving hooks from pool provider to hooks folder. Disabling withdraw button if time lock has not expired
