+++ src/components/Farming/Approve/index.js
@@ -1,6 +1,6 @@
 import React, { useState, useRef } from 'react'
 import { Button, TransactionProgress } from '@1hive/1hive-ui'
-import { useApprove } from '../../../providers/Poolprovider'
+import { useApprove } from '../../../hooks/useApprove'
 import { getNetworkConfig } from '../../../networks'
 
 const Approve = props => {
@@ -29,14 +29,6 @@ const Approve = props => {
 
   return (
     <>
-      <Button
-        css={`
-          background: linear-gradient(90deg, #aaf5d4, #7ce0d6);
-        `}
-        label="Approve"
-        wide
-        onClick={handleApprove}
-      />
       <TransactionProgress
         transactionHashUrl={network.txUrl + txHash}
         progress={0.3}
@@ -45,6 +37,15 @@ const Approve = props => {
         onClose={() => setVisible(false)}
         opener={opener}
         slow
+      />
+      <Button
+        css={`
+          background: linear-gradient(90deg, #aaf5d4, #7ce0d6);
+        `}
+        label="Approve"
+        wide
+        onClick={handleApprove}
+        ref={opener}
       />
     </>
   )
