--- src/components/Farming/Deposit/deposit.style.js
@@ -1,11 +1,0 @@
-import styled from 'styled-components'
-
-const TransactionProgress = styled.div`
-  position: absolute;
-  right: 25px !important;
-`
-const Styled = {
-  TransactionProgress,
-}
-
-export default Styled
