+++ src/components/Farming/DepositModal/index.js
@@ -3,7 +3,7 @@ import { Modal, GU } from '@1hive/1hive-ui'
 import { ethers } from 'ethers'
 import SliderComponent from '../SliderComponent'
 import { getKnownTokenImg } from '../../../utils/known-tokens'
-import { useCheckApprovedToken } from '../../../providers/Poolprovider'
+import { useCheckApprovedToken } from '../../../hooks/useCheckApproved'
 import Approved from '../Approve'
 import Deposit from '../Deposit'
 import multiplier from '../../../assets/multiplier.svg'
