+++ src/hooks/useHarvest.js
@@ -1,0 +1,5 @@
+export function useHarvest(id) {
+  return () => {
+    // TODO: Add harvest function here
+  }
+}
